+++ x/emissions/keeper/actor_utils/worker.go
@@ -203,6 +203,8 @@ func ProcessAndStoreNetworkInferences(
 		return errorsmod.Wrap(err, "failed to insert outlier resistant network inference")
 	}
 
+	types.EmitNewOutlierResistantNetworkInferencesEvent(ctx, topicId, blockHeight, *outlierResistantNetworkInferencesResult.NetworkInferences)
+
 	return nil
 }